Add Hitable concept and a Sphere class
Encapsualate the sphere logic into a sphere class that conforms to the
Hitable abstract class.

Add a HitableList class which also conforms to the Hitable abstract
class. HitableList is a structure that holds a list of Hitable objects,
which test hit against all the Hitable objects returning only the
closeest hit point to the camera.